update summarize_NoLoCo to improve synapse detection
Subtract local background (using a median filter) when identifying synapses from activity image.
The root problem here is that different regions of the image are recorded different numbers of times. outputting the 'freshness' images when getting frames from the dataFile or saving the per-pixel framerate in metadata would address this. The median filter is a workaround but may be useful regardless.